+++ docs/CHANGELOG.md
@@ -64,6 +64,7 @@
     * `Implement RaspberryJuice API` (Enabled By Default)
     * `Fix HUD When Spectating Other Players` (Enabled By Default)
     * `Fix Crash When Spectated Entity Is Removed` (Enabled By Default)
+    * `Fix Torch Placement` (Enabled)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`